requirements.txt
webdriver_manager
selenium==4.5.0
beautifulsoup4==4.12.3
PyYAML==6.0.1
pymongo